reid/scripts/triplet_reid/requirements.txt
tqdm==4.28.1
pymongo==3.7.2
setuptools==28.8.0
requests==2.18.4
mock==2.0.0
ipdb==0.11
py_cpuinfo==4.0.0
pandas==0.22.0
wrapt==1.10.11
h5py==2.9.0
scipy==1.2.0
torchvision==0.2.0
torch==1.0.1.post2
objgraph==3.4.0
matplotlib==2.1.0
pycocotools==2.0.0
munch==2.3.2
imgaug==0.2.8
pytest==3.9.3
tensorflow_gpu==1.12.0
opencv_contrib_python==3.4.0.12
colorama==0.4.0
numpy==1.16.2
jsonpickle==0.9.6
docopt==0.6.2
Pillow==5.4.1
GitPython==2.1.11
hashfs==0.7.1
Mako==1.0.7
sphinx_rtd_theme==0.4.3
SQLAlchemy==1.2.18
telegram==0.0.1
tensorboardX==1.6
tensorflow==1.13.1
tinydb==3.12.2
tinydb_serialization==1.0.4
sacred==0.7.4
